pyproject.toml
[tool.poetry]
name = "django-simple"
version = "0.1.0"
description = "A simple Django project"
authors = ["Ryan Schuetzler <ryan@schuetzler.net>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
django-crispy-forms = "^1.14.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
